Extract generated LOD1 roof surface factory
