+++ src/PlateauResoniteLink/Application/Importing/LocalCityGmlObjectProjection.cs
@@ -1054,7 +1054,7 @@ internal static class LocalCityGmlObjectProjection
         }
 
         global::PlateauResoniteLink.Application.Importing.ParsedSurface[] generatedSurfaces =
-            CreateGeneratedLod1RoofSurfaces(resolvedFootprint, roofShape);
+            GeneratedLod1RoofSurfaceFactory.Create(resolvedFootprint, roofShape);
         if (generatedSurfaces.Length == 0)
         {
             return cityObject;
@@ -1202,265 +1202,6 @@ internal static class LocalCityGmlObjectProjection
     {
         double scale = Math.Max(Math.Max(Math.Abs(left), Math.Abs(right)), 1.0);
         return Math.Abs(left - right) <= scale * relativeTolerance;
-    }
-
-    private static global::PlateauResoniteLink.Application.Importing.ParsedSurface[] CreateGeneratedLod1RoofSurfaces(
-        Lod1RoofFootprint footprint,
-        GeneratedLod1RoofShape shape)
-    {
-        double rise = ComputeGeneratedRoofRiseMeters(footprint);
-        return shape switch
-        {
-            GeneratedLod1RoofShape.Shed => CreateShedRoofSurfaces(footprint, rise),
-            GeneratedLod1RoofShape.Gable => CreateGableRoofSurfaces(footprint, rise),
-            GeneratedLod1RoofShape.Hip => CreateHipRoofSurfaces(footprint, rise),
-            _ => [],
-        };
-    }
-
-    private static double ComputeGeneratedRoofRiseMeters(Lod1RoofFootprint footprint)
-    {
-        double contextualLimit = Math.Max(0.6, footprint.GeometryHeightMeters * 0.18);
-        return Math.Clamp(Math.Min(footprint.WidthMeters * 0.28, contextualLimit), 0.4, 2.2);
-    }
-
-    private static global::PlateauResoniteLink.Application.Importing.ParsedSurface[] CreateShedRoofSurfaces(
-        Lod1RoofFootprint footprint,
-        double rise)
-    {
-        global::PlateauResoniteLink.Application.Importing.GeodeticPoint[] c = footprint.Corners;
-        bool firstLong = footprint.FirstEdgeIsLongAxis;
-        global::PlateauResoniteLink.Application.Importing.GeodeticPoint[] highEdge = firstLong ? [Elevate(c[2], rise), Elevate(c[3], rise)] : [Elevate(c[1], rise), Elevate(c[2], rise)];
-        global::PlateauResoniteLink.Application.Importing.GeodeticPoint[] roof = firstLong
-            ? [c[0], c[1], highEdge[0], highEdge[1]]
-            : [c[0], highEdge[0], highEdge[1], c[3]];
-
-        List<global::PlateauResoniteLink.Application.Importing.ParsedSurface> surfaces =
-        [
-            CreateGeneratedRoofSurface(footprint, "shed-roof", ParsedSurfaceSemantic.Roof, roof),
-        ];
-        if (firstLong)
-        {
-            surfaces.Add(CreateGeneratedRoofSurface(footprint, "shed-high-wall", ParsedSurfaceSemantic.Wall, [c[3], c[2], highEdge[0], highEdge[1]]));
-            surfaces.Add(CreateGeneratedRoofSurface(footprint, "shed-side-wall-a", ParsedSurfaceSemantic.Wall, [c[1], c[2], highEdge[0]]));
-            surfaces.Add(CreateGeneratedRoofSurface(footprint, "shed-side-wall-b", ParsedSurfaceSemantic.Wall, [c[0], highEdge[1], c[3]]));
-        }
-        else
-        {
-            surfaces.Add(CreateGeneratedRoofSurface(footprint, "shed-high-wall", ParsedSurfaceSemantic.Wall, [c[1], c[2], highEdge[1], highEdge[0]]));
-            surfaces.Add(CreateGeneratedRoofSurface(footprint, "shed-side-wall-a", ParsedSurfaceSemantic.Wall, [c[0], c[1], highEdge[0]]));
-            surfaces.Add(CreateGeneratedRoofSurface(footprint, "shed-side-wall-b", ParsedSurfaceSemantic.Wall, [c[3], highEdge[1], c[2]]));
-        }
-
-        return surfaces.ToArray();
-    }
-
-    private static global::PlateauResoniteLink.Application.Importing.ParsedSurface[] CreateGableRoofSurfaces(
-        Lod1RoofFootprint footprint,
-        double rise)
-    {
-        global::PlateauResoniteLink.Application.Importing.GeodeticPoint[] c = footprint.Corners;
-        bool firstLong = footprint.FirstEdgeIsLongAxis;
-        global::PlateauResoniteLink.Application.Importing.GeodeticPoint ridge0;
-        global::PlateauResoniteLink.Application.Importing.GeodeticPoint ridge1;
-        if (firstLong)
-        {
-            ridge0 = Elevate(Lerp(c[0], c[3], 0.5), rise);
-            ridge1 = Elevate(Lerp(c[1], c[2], 0.5), rise);
-            return
-            [
-                CreateGeneratedRoofSurface(footprint, "gable-roof-a", ParsedSurfaceSemantic.Roof, [c[0], c[1], ridge1, ridge0]),
-                CreateGeneratedRoofSurface(footprint, "gable-roof-b", ParsedSurfaceSemantic.Roof, [c[3], ridge0, ridge1, c[2]]),
-                CreateGeneratedRoofSurface(footprint, "gable-wall-a", ParsedSurfaceSemantic.Wall, [c[0], ridge0, c[3]]),
-                CreateGeneratedRoofSurface(footprint, "gable-wall-b", ParsedSurfaceSemantic.Wall, [c[1], c[2], ridge1]),
-            ];
-        }
-
-        ridge0 = Elevate(Lerp(c[0], c[1], 0.5), rise);
-        ridge1 = Elevate(Lerp(c[3], c[2], 0.5), rise);
-        return
-        [
-            CreateGeneratedRoofSurface(footprint, "gable-roof-a", ParsedSurfaceSemantic.Roof, [c[0], ridge0, ridge1, c[3]]),
-            CreateGeneratedRoofSurface(footprint, "gable-roof-b", ParsedSurfaceSemantic.Roof, [ridge0, c[1], c[2], ridge1]),
-            CreateGeneratedRoofSurface(footprint, "gable-wall-a", ParsedSurfaceSemantic.Wall, [c[0], c[1], ridge0]),
-            CreateGeneratedRoofSurface(footprint, "gable-wall-b", ParsedSurfaceSemantic.Wall, [c[3], ridge1, c[2]]),
-        ];
-    }
-
-    private static global::PlateauResoniteLink.Application.Importing.ParsedSurface[] CreateHipRoofSurfaces(
-        Lod1RoofFootprint footprint,
-        double rise)
-    {
-        global::PlateauResoniteLink.Application.Importing.GeodeticPoint[] c = footprint.Corners;
-        if (footprint.FirstEdgeIsLongAxis)
-        {
-            global::PlateauResoniteLink.Application.Importing.GeodeticPoint leftMid = Lerp(c[0], c[3], 0.5);
-            global::PlateauResoniteLink.Application.Importing.GeodeticPoint rightMid = Lerp(c[1], c[2], 0.5);
-            global::PlateauResoniteLink.Application.Importing.GeodeticPoint longRidge0 = Elevate(Lerp(leftMid, rightMid, 0.25), rise);
-            global::PlateauResoniteLink.Application.Importing.GeodeticPoint longRidge1 = Elevate(Lerp(leftMid, rightMid, 0.75), rise);
-            return
-            [
-                CreateGeneratedRoofSurface(footprint, "hip-roof-a", ParsedSurfaceSemantic.Roof, [c[0], c[1], longRidge1, longRidge0]),
-                CreateGeneratedRoofSurface(footprint, "hip-roof-b", ParsedSurfaceSemantic.Roof, [c[3], longRidge0, longRidge1, c[2]]),
-                CreateGeneratedRoofSurface(footprint, "hip-roof-c", ParsedSurfaceSemantic.Roof, [c[0], longRidge0, c[3]]),
-                CreateGeneratedRoofSurface(footprint, "hip-roof-d", ParsedSurfaceSemantic.Roof, [c[1], c[2], longRidge1]),
-            ];
-        }
-
-        global::PlateauResoniteLink.Application.Importing.GeodeticPoint bottomMid = Lerp(c[0], c[1], 0.5);
-        global::PlateauResoniteLink.Application.Importing.GeodeticPoint topMid = Lerp(c[3], c[2], 0.5);
-        global::PlateauResoniteLink.Application.Importing.GeodeticPoint shortRidge0 = Elevate(Lerp(bottomMid, topMid, 0.25), rise);
-        global::PlateauResoniteLink.Application.Importing.GeodeticPoint shortRidge1 = Elevate(Lerp(bottomMid, topMid, 0.75), rise);
-        return
-        [
-            CreateGeneratedRoofSurface(footprint, "hip-roof-a", ParsedSurfaceSemantic.Roof, [c[0], shortRidge0, shortRidge1, c[3]]),
-            CreateGeneratedRoofSurface(footprint, "hip-roof-b", ParsedSurfaceSemantic.Roof, [shortRidge0, c[1], c[2], shortRidge1]),
-            CreateGeneratedRoofSurface(footprint, "hip-roof-c", ParsedSurfaceSemantic.Roof, [c[0], c[1], shortRidge0]),
-            CreateGeneratedRoofSurface(footprint, "hip-roof-d", ParsedSurfaceSemantic.Roof, [c[3], shortRidge1, c[2]]),
-        ];
-    }
-
-    private static global::PlateauResoniteLink.Application.Importing.ParsedSurface CreateGeneratedRoofSurface(
-        Lod1RoofFootprint footprint,
-        string suffix,
-        ParsedSurfaceSemantic semantic,
-        global::PlateauResoniteLink.Application.Importing.GeodeticPoint[] vertices)
-    {
-        string polygonId = $"{footprint.TopSurface.PolygonId}_generated_{suffix}";
-        global::PlateauResoniteLink.Application.Importing.GeodeticPoint[] orientedVertices =
-            semantic == ParsedSurfaceSemantic.Wall
-                ? OrientGeneratedWallVerticesForOutwardMeshFaces(footprint, vertices)
-                : semantic == ParsedSurfaceSemantic.Roof
-                ? OrientGeneratedRoofVerticesForUpwardMeshFaces(footprint, vertices)
-                : vertices;
-        global::PlateauResoniteLink.Application.Importing.GeodeticPoint[] closedVertices = [.. orientedVertices, orientedVertices[0]];
-        return new global::PlateauResoniteLink.Application.Importing.ParsedSurface(
-            polygonId,
-            (global::PlateauResoniteLink.Application.Importing.ParsedSurfaceSemantic)semantic,
-            new global::PlateauResoniteLink.Application.Importing.ParsedRing(
-                $"{polygonId}-ring",
-                closedVertices,
-                UVs: null),
-            InteriorRings: [],
-            footprint.TopSurface.BaseColor,
-            TexturePayload: null,
-            UsesGeneratedDemTexture: false,
-            footprint.TopSurface.OpticalProperties);
-    }
-
-    private static global::PlateauResoniteLink.Application.Importing.GeodeticPoint[] OrientGeneratedWallVerticesForOutwardMeshFaces(
-        Lod1RoofFootprint footprint,
-        global::PlateauResoniteLink.Application.Importing.GeodeticPoint[] vertices)
-    {
-        if (vertices.Length < 3)
-        {
-            return vertices;
-        }
-
-        double referenceLatitude = footprint.Corners.Average(static point => point.Latitude);
-        double referenceLongitude = footprint.Corners.Average(static point => point.Longitude);
-        Float3[] footprintPositions = footprint.Corners
-            .Select(point => CreateApproximateHorizontalPosition(point, referenceLatitude, referenceLongitude))
-            .ToArray();
-        Float3[] wallPositions = vertices
-            .Select(point => CreateApproximateHorizontalPosition(point, referenceLatitude, referenceLongitude))
-            .ToArray();
-        Float3? normal = ComputePolygonNormal(wallPositions);
-        if (normal is null)
-        {
-            return vertices;
-        }
-
-        Float3 footprintCenter = AveragePosition(footprintPositions);
-        Float3 wallCenter = AveragePosition(wallPositions);
-        Float3 outwardDirection = new(wallCenter.X - footprintCenter.X, 0.0, wallCenter.Z - footprintCenter.Z);
-        Float3 horizontalNormal = new(normal.X, 0.0, normal.Z);
-        // Surface tessellation emits Resonite triangle winding opposite to the parsed polygon normal.
-        // Generated walls therefore need an inward parsed normal so the emitted mesh face points outward.
-        if (Dot(horizontalNormal, outwardDirection) <= 0.0)
-        {
-            return vertices;
-        }
-
-        global::PlateauResoniteLink.Application.Importing.GeodeticPoint[] reversed = vertices.ToArray();
-        Array.Reverse(reversed);
-        return reversed;
-    }
-
-    private static global::PlateauResoniteLink.Application.Importing.GeodeticPoint[] OrientGeneratedRoofVerticesForUpwardMeshFaces(
-        Lod1RoofFootprint footprint,
-        global::PlateauResoniteLink.Application.Importing.GeodeticPoint[] vertices)
-    {
-        if (vertices.Length < 3)
-        {
-            return vertices;
-        }
-
-        double referenceLatitude = footprint.Corners.Average(static point => point.Latitude);
-        double referenceLongitude = footprint.Corners.Average(static point => point.Longitude);
-        Float3[] roofPositions = vertices
-            .Select(point => CreateApproximateHorizontalPosition(point, referenceLatitude, referenceLongitude))
-            .ToArray();
-        Float3? normal = ComputePolygonNormal(roofPositions);
-        if (normal is null)
-        {
-            return vertices;
-        }
-
-        // Surface tessellation emits Resonite triangle winding opposite to the parsed polygon normal.
-        // Generated roofs therefore need a downward parsed normal so the emitted mesh face points upward.
-        if (normal.Y <= 0.0)
-        {
-            return vertices;
-        }
-
-        global::PlateauResoniteLink.Application.Importing.GeodeticPoint[] reversed = vertices.ToArray();
-        Array.Reverse(reversed);
-        return reversed;
-    }
-
-    private static Float3 CreateApproximateHorizontalPosition(
-        global::PlateauResoniteLink.Application.Importing.GeodeticPoint point,
-        double referenceLatitude,
-        double referenceLongitude)
-    {
-        const double metersPerLatitudeDegree = 111_320.0;
-        double metersPerLongitudeDegree = metersPerLatitudeDegree * Math.Cos(referenceLatitude * (Math.PI / 180.0));
-        return new Float3(
-            (point.Longitude - referenceLongitude) * metersPerLongitudeDegree,
-            point.Altitude,
-            (point.Latitude - referenceLatitude) * metersPerLatitudeDegree);
-    }
-
-    private static Float3 AveragePosition(IReadOnlyList<Float3> positions)
-    {
-        return new Float3(
-            positions.Average(static position => position.X),
-            positions.Average(static position => position.Y),
-            positions.Average(static position => position.Z));
-    }
-
-    private static GeodeticPoint Elevate(GeodeticPoint point, double rise)
-    {
-        return point with { Altitude = point.Altitude + rise };
-    }
-
-    private static global::PlateauResoniteLink.Application.Importing.GeodeticPoint Elevate(
-        global::PlateauResoniteLink.Application.Importing.GeodeticPoint point,
-        double rise)
-    {
-        return point with { Altitude = point.Altitude + rise };
-    }
-
-    private static global::PlateauResoniteLink.Application.Importing.GeodeticPoint Lerp(
-        global::PlateauResoniteLink.Application.Importing.GeodeticPoint source,
-        global::PlateauResoniteLink.Application.Importing.GeodeticPoint target,
-        double ratio)
-    {
-        return new global::PlateauResoniteLink.Application.Importing.GeodeticPoint(
-            source.Latitude + ((target.Latitude - source.Latitude) * ratio),
-            source.Longitude + ((target.Longitude - source.Longitude) * ratio),
-            source.Altitude + ((target.Altitude - source.Altitude) * ratio));
     }
 
     private static GeodeticPoint ResolveProjectionCityObjectOrigin(ParsedCityObject cityObject)
@@ -4991,15 +4732,6 @@ internal static class LocalCityGmlObjectProjection
         public IEnumerable<GeodeticPoint> Vertices =>
             ExteriorRing.Vertices.Concat(InteriorRings.SelectMany(static ring => ring.Vertices));
     }
-
-    private sealed record Lod1RoofFootprint(
-        global::PlateauResoniteLink.Application.Importing.ParsedSurface TopSurface,
-        global::PlateauResoniteLink.Application.Importing.GeodeticPoint[] Corners,
-        double LengthMeters,
-        double WidthMeters,
-        double GeometryHeightMeters,
-        BuildingAttributeContext Attributes,
-        bool FirstEdgeIsLongAxis);
 
     private sealed record TessellatedVertex(
         Float3 Position,
